ran with flags

+++ tests/test_reverse_iterator.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   test_reverse_iterator.cpp                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kmilchev <kmilchev@student.42wolfsburg.de> +#+  +:+       +#+        */
+/*   By: zangelis <zangelis@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/21 17:57:04 by kmilchev          #+#    #+#             */
-/*   Updated: 2022/09/16 18:18:30 by kmilchev         ###   ########.fr       */
+/*   Updated: 2022/10/03 00:45:49 by zangelis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ int main(void)
 	r = nums.rbegin();
 
 	test_name("Testing the reverse_iterator. Checking if values are in reverse order");
-	for(int i = 0; i < nums.size(); i++)
-		assert(nums.size() - i - 1 == *(r  + i));// 9 8 7 6 5 4 3 2 1 0
+	for(size_t i = 0; i < nums.size(); i++)
+		assert(int(nums.size() - i - 1) == *(r  + i));// 9 8 7 6 5 4 3 2 1 0
 	std::cout << COLOUR_DEFAULT;
 }
